+++ backend/src/baserow/contrib/database/api/views/serializers.py
@@ -91,9 +91,16 @@ class FieldOptionsField(serializers.Field):
 
 
 class ViewFilterSerializer(serializers.ModelSerializer):
+    preload_values = serializers.DictField(
+        help_text="Can contain unique preloaded values per filter. This is for "
+        "example used by the `link_row_has` filter to communicate the display name if "
+        "a value is provided.",
+        read_only=True,
+    )
+
     class Meta:
         model = ViewFilter
-        fields = ("id", "view", "field", "type", "value")
+        fields = ("id", "view", "field", "type", "value", "preload_values")
         extra_kwargs = {"id": {"read_only": True}}