+++ web-frontend/modules/core/utils/string.js
@@ -87,3 +87,7 @@ export const isSecureURL = (str) => {
 export const escapeRegExp = (string) => {
   return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
 }
+
+export const isNumeric = (value) => {
+  return /^-?\d+$/.test(value)
+}