+++ web-frontend/modules/database/components/view/ViewFilterTypeSelectOptions.vue
@@ -10,41 +10,16 @@
 
 <script>
 import FieldSingleSelectDropdown from '@baserow/modules/database/components/field/FieldSingleSelectDropdown'
+import viewFilter from '@baserow/modules/database/mixins/viewFilter'
 
 export default {
   name: 'ViewFilterTypeSelectOptions',
   components: { FieldSingleSelectDropdown },
-  props: {
-    value: {
-      type: String,
-      required: true,
-    },
-    fieldId: {
-      type: Number,
-      required: true,
-    },
-    primary: {
-      type: Object,
-      required: true,
-    },
-    fields: {
-      type: Array,
-      required: true,
-    },
-    readOnly: {
-      type: Boolean,
-      required: true,
-    },
-  },
+  mixins: [viewFilter],
   computed: {
     copy() {
-      const value = this.value
+      const value = this.filter.value
       return value === '' ? null : parseInt(value) || null
-    },
-    field() {
-      return this.primary.id === this.fieldId
-        ? this.primary
-        : this.fields.find((f) => f.id === this.fieldId)
     },
   },
   methods: {