+++ changelog.md
@@ -9,6 +9,7 @@
   database has been removed entirely to prevent data integrity issues.
 * Fixed bug where the currently selected dropdown item is out of view from the dropdown
   window when scrolling with the arrow keys.
+* Introduced link row field has row filter.
 * Made the form view compatible with importing and exporting.
 * Made it possible to use the "F2"-Key to edit a cell without clearing the cell content.
 * Added password validation to password reset page.